@@ -1,6 +1,6 @@
 {
   "name": "skeleton",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "license": "UNLICENSED",
   "scripts": {
     "prebuild": "rm -rf dist",
@@ -10,13 +10,13 @@
     "test": "jest --colors"
   },
   "dependencies": {
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "devDependencies": {
-    "@adhamu/zero": "^2.0.8",
+    "@adhamu/zero": "^4.0.0",
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
+    "@testing-library/react": "^13.0.0",
     "@testing-library/react-hooks": "^7.0.0",
     "@types/jest": "^27.4.0",
     "@types/react": "^17.0.1",
