@@ -1,7 +1,5 @@
-import * as React from 'react'
-
-import * as ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 
 import App from './App'
 
-ReactDOM.render(<App />, document.getElementById('app'))
+createRoot(document.getElementById('app') as HTMLElement).render(<App />)
